feat: <1단계> 피드백 반영
- StringCalculator 클래스 기능 분리
